Move ShouldRequireCustomsForm use case to domain package
